add esm build that autogens .d.ts files from jsdoc using tsc
