+++ package.json
@@ -24,6 +24,7 @@
     "scripts": {
         "build:examples": "literally -c literally.config.js",
         "build:rollup": "rollup -c rollup.config.js",
+        "build:ts": "tsc",
         "build": "npm-run-all -p build:*",
         "clean": "rimraf dist",
         "docs": "jsdoc2md src/js/index.js --separators > api.md",
@@ -33,6 +34,7 @@
         "test": "jest --noStackTrace",
         "watch:examples": "literally --watch",
         "watch:rollup": "rollup -c rollup.config.js --watch",
+        "watch:ts": "tsc --watch",
         "watch": "npm-run-all -p watch:*"
     },
     "publishConfig": {
@@ -77,6 +79,7 @@
         "rollup-plugin-terser": "^6.1.0",
         "source-map-explorer": "^2.4.2",
         "source-map-loader": "^0.2.4",
-        "superstore-arrow": "^1.0.0"
+        "superstore-arrow": "^1.0.0",
+        "typescript": "^3.9.5"
     }
 }
