+++ tsconfig.json
@@ -1,0 +1,16 @@
+{
+  "$schema": "http://json.schemastore.org/tsconfig",
+  "compilerOptions": {
+    "allowJs": true,
+    "checkJs": false,
+    "declaration": true,
+    "experimentalDecorators": true,
+    "module": "esnext",
+    "moduleResolution": "node",
+    "target": "esnext",
+
+    "outDir": "dist/esm",
+    "rootDir": "src/js"
+  },
+  "include": ["src/**/*"]
+}
